+++ src/mass_spectrometr/GUI/Volt_engine_fast.java
@@ -28,17 +28,17 @@ public class Volt_engine_fast extends Volt_engine {
 	}
 
 	@Override
-	float get_start_V() {
+	double get_start_V() {
 		return calc_float_from_int(Run.prog.start_V_cyclic);
 	}
 
 	@Override
-	float get_stop_V() {
+	double get_stop_V() {
 		return calc_float_from_int(Run.prog.stop_V_cyclic);
 	}
 
 	@Override
-	float get_step_V() {
+	double get_step_V() {
 		return Run.prog.step_V_cyclic;
 	}
 
@@ -48,18 +48,18 @@ public class Volt_engine_fast extends Volt_engine {
 	}
 
 	@Override
-	void set_start_V(float v) {
+	void set_start_V(double v) {
 		Run.prog.start_V_cyclic = calc_int_from_float(v);
 	}
 
 	@Override
-	void set_stop_V(float v) {
+	void set_stop_V(double v) {
 		Run.prog.stop_V_cyclic = calc_int_from_float(v);
 
 	}
 
 	@Override
-	void set_step_V(float v) {
+	void set_step_V(double v) {
 		Run.prog.step_V_cyclic = calc_step(v);
 
 	}
