+++ src/mass_spectrometr/GUI/Volt_engine_long.java
@@ -5,17 +5,17 @@ import mass_spectrometr.Run;
 public class Volt_engine_long extends Volt_engine{
 
 	@Override
-	float get_start_V() {
+	double get_start_V() {
 		return calc_float_from_int(Run.prog.start_V);
 	}
 
 	@Override
-	float get_stop_V() {
+	double get_stop_V() {
 		return calc_float_from_int(Run.prog.stop_V);
 	}
 
 	@Override
-	float get_step_V() {
+	double get_step_V() {
 		return Run.prog.step_V;
 	}
 
@@ -25,18 +25,18 @@ public class Volt_engine_long extends Volt_engine{
 	}
 
 	@Override
-	void set_start_V(float v) {
+	void set_start_V(double v) {
 		Run.prog.start_V = calc_int_from_float(v);
 	}
 
 	@Override
-	void set_stop_V(float v) {
+	void set_stop_V(double v) {
 		Run.prog.stop_V = calc_int_from_float(v);
 		
 	}
 
 	@Override
-	void set_step_V(float v) {
+	void set_step_V(double v) {
 		Run.prog.step_V = calc_step(v);
 		
 	}
